+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_help.py
@@ -46,24 +46,29 @@ helps['webapp auth update'] = """
 
 helps['webapp identity assign'] = """
     type: command
-    short-summary: (PREVIEW) assign managed service identity to the webapp
+    short-summary: assign or disable managed service identity to the webapp
     examples:
         - name: assign local identity and assign a reader role to the current resource group.
           text: >
             az webapp identity assign -g MyResourceGroup -n MyUniqueApp --role reader --scope /subscriptions/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/MyResourceGroup
-        - name: disable the identity when there is need.
+        - name: enable identity for the webapp.
           text: >
-            az webapp config appsettings set -g MyResourceGroup -n MyUniqueApp --settings WEBSITE_DISABLE_MSI=true
+            az webapp identity assign -g MyResourceGroup -n MyUniqueApp
 """
 
 helps['webapp identity'] = """
     type: group
-    short-summary: (PREVIEW) manage webapp's managed service identity
+    short-summary: manage webapp's managed service identity
 """
 
 helps['webapp identity show'] = """
     type: command
-    short-summary: (PREVIEW) display webapp's managed service identity
+    short-summary: display webapp's managed service identity
+"""
+
+helps['webapp identity remove'] = """
+    type: command
+    short-summary: Disable webapp's managed service identity
 """
 
 helps['functionapp identity'] = helps['webapp identity'].replace('webapp', 'functionapp')
@@ -71,6 +76,8 @@ helps['functionapp identity'] = helps['webapp identity'].replace('webapp', 'func
 helps['functionapp identity assign'] = helps['webapp identity assign'].replace('webapp', 'functionapp')
 
 helps['functionapp identity show'] = helps['webapp identity show'].replace('webapp', 'functionapp')
+
+helps['functionapp identity remove'] = helps['webapp identity remove'].replace('webapp', 'functionapp')
 
 helps['webapp config'] = """
 type: group